choosing clients
There were lots of leads, where it made no sense (to small, not a cultural fit etc) to work for the client, or there was no resource. Moreover, I've been asked to help with leads, which I had no info about. I can only fulfill the purpose of this role, if we add these