naming convention
Tension: I've been asked several time, but could not give a proper and transparent answer, which is more of an issue as we are growing. Moreover, Cassus told me at least twice, that he hates me using the term CEO.
Solution: taking the best practice of consulting (like big4) and law firms, which make the seniors 'partners' with a buy-in option. The term Partner fits both with common sense business language and with holacracy. Managing partner is not perfect, but is my best guess, and is better than CEO. Lets be realistic, as we still need something CXO-ish for the external world, outside of our bubble... #communications