Clarified FTE definition
Tension: With the previous FTE definition, I did not know how to correctly interpret it when it came down to being a 'full-time' dev on Streamit fulfilling a daily-rate contract. This change clarifies that the 80/20 FTE split is an internal Lab.Coop guidelines, not a contractual specification (at least, not so far), that must be explicitly communicated with the clients by the client teams.